+++ cmd/data-node/commands/dehistory/dehistory.go
@@ -2,8 +2,12 @@ package dehistory
 
 import (
 	"context"
+	"fmt"
 
 	"code.vegaprotocol.io/vega/datanode/dehistory"
+	vgjson "code.vegaprotocol.io/vega/libs/json"
+	"code.vegaprotocol.io/vega/logging"
+	"code.vegaprotocol.io/vega/paths"
 
 	"github.com/jessevdk/go-flags"
 	"google.golang.org/grpc"
@@ -54,4 +58,44 @@ func datanodeLive(cfg config.Config) bool {
 
 	_, err = client.Ping(context.Background(), &v2.PingRequest{})
 	return err == nil
+}
+
+// getConfig figures out where to read a config file from, reads it, and then applies any extra
+// modifications on top of that.
+//
+// This is working around a bit of awkwardness in that the config supplied by go-flags is a blank
+// config updated with command line flags. There is not enough information in it to apply an
+// 'overlay' to a config read from a file, because it is not possible for us to tell if someone
+// is trying to override a value back to it's 'zero' value. (e.g. --something.enabled=false gives
+// the same go-flags structure as no argument at all).
+func fixConfig(config *config.Config, vegaPaths paths.Paths) error {
+	configFilePath, err := vegaPaths.CreateConfigPathFor(paths.DataNodeDefaultConfigFile)
+	if err != nil {
+		return fmt.Errorf("couldn't get path for %s: %w", paths.DataNodeDefaultConfigFile, err)
+	}
+
+	// Read config from file
+	err = paths.ReadStructuredFile(configFilePath, config)
+	if err != nil {
+		return fmt.Errorf("failed to read config:%w", err)
+	}
+
+	// Apply command-line flags on top
+	_, err = flags.NewParser(config, flags.Default|flags.IgnoreUnknown).Parse()
+	if err != nil {
+		return fmt.Errorf("failed to parse args:%w", err)
+	}
+	return nil
+}
+
+func handleErr(log *logging.Logger, outputJSON bool, msg string, err error) {
+	if outputJSON {
+		_ = vgjson.Print(struct {
+			Error string `json:"error"`
+		}{
+			Error: err.Error(),
+		})
+	} else {
+		log.Error(msg, logging.Error(err))
+	}
 }